requirements.txt
streamlit
numpy
pypdf
langchain
langchain-openai
langchain-text-splitters
langchain-community
python-dotenv
chromadb
python-docx
python-pptx
sentence-transformers